refactor(badges): improve BadgesGrid component structure and maintainability
