+++ src/components/Dashboard/StatsSection/StatsOverview/BadgesGrid/animations/index.ts
@@ -1,0 +1,14 @@
+export const container = {
+  hidden: { opacity: 0 },
+  show: {
+    opacity: 1,
+    transition: {
+      staggerChildren: 0.1
+    }
+  }
+};
+
+export const item = {
+  hidden: { opacity: 0, y: 20 },
+  show: { opacity: 1, y: 0 }
+};
