+++ src/components/Dashboard/StatsSection/StatsOverview/BadgesGrid/hooks/useFormatDate.ts
@@ -1,0 +1,17 @@
+export const useFormatDate = () => {
+  return (dateString: string) => {
+    try {
+      const date = new Date(dateString);
+      if (isNaN(date.getTime())) {
+        return 'Data nieznana';
+      }
+      return date.toLocaleDateString('pl-PL', {
+        year: 'numeric',
+        month: 'short',
+        day: 'numeric'
+      });
+    } catch {
+      return 'Data nieznana';
+    }
+  };
+};
